Move extensions install test into same file so things run in sync instead of parallel
